Make all() and num_range() return vecs
`all()` and `num_range()` now return a `Vec` of `CountryCode`s, and not an
`Option` of a `Vec`. This was worthless because instead of a `Some` if there
was at least one entry or a `None` if there were none, a `len()` check on the
`Vec` can just be performed.